Crate google_dialogflow2

This documentation was generated from Dialogflow crate version 1.0.9+20190629, where 20190629 is the exact revision of the dialogflow:v2 schema built by the mako code generator v1.0.9.

Everything else about the Dialogflow v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

projects

agent, agent entity types batch delete, agent entity types batch update, agent entity types create, agent entity types delete, agent entity types entities batch create, agent entity types entities batch delete, agent entity types entities batch update, agent entity types get, agent entity types list, agent entity types patch, agent export, agent import, agent intents batch delete, agent intents batch update, agent intents create, agent intents delete, agent intents get, agent intents list, agent intents patch, agent restore, agent search, agent sessions contexts create, agent sessions contexts delete, agent sessions contexts get, agent sessions contexts list, agent sessions contexts patch, agent sessions delete contexts, agent sessions detect intent, agent sessions entity types create, agent sessions entity types delete, agent sessions entity types get, agent sessions entity types list, agent sessions entity types patch, agent train, delete agent, get agent and operations get

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.projects().agent_restore(...).doit()
let r = hub.projects().agent_intents_batch_delete(...).doit()
let r = hub.projects().agent_entity_types_batch_delete(...).doit()
let r = hub.projects().agent_import(...).doit()
let r = hub.projects().agent_export(...).doit()
let r = hub.projects().agent_entity_types_batch_update(...).doit()
let r = hub.projects().agent_train(...).doit()
let r = hub.projects().agent_intents_batch_update(...).doit()
let r = hub.projects().agent_entity_types_entities_batch_update(...).doit()
let r = hub.projects().agent_entity_types_entities_batch_delete(...).doit()
let r = hub.projects().operations_get(...).doit()
let r = hub.projects().agent_entity_types_entities_batch_create(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-dialogflow2 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_dialogflow2 as dialogflow2;
use dialogflow2::GoogleCloudDialogflowV2RestoreAgentRequest;
use dialogflow2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dialogflow2::Dialogflow;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dialogflow::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = GoogleCloudDialogflowV2RestoreAgentRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().agent_restore(req, "parent")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

Chunk

ContentRange
Implements the Content-Range header, for serialization only

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Dialogflow
Central instance to access all Dialogflow related resource activities

DummyNetworkStream

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

GoogleCloudDialogflowV2DetectIntentRequest
The request to detect user's intent.

GoogleCloudDialogflowV2IntentMessage
Corresponds to the Response field in the Dialogflow console.

GoogleCloudDialogflowV2IntentMessageSimpleResponses
The collection of simple response candidates. This message in QueryResult.fulfillment_messages and WebhookResponse.fulfillment_messages should contain only one SimpleResponse.

GoogleCloudDialogflowV2QueryInput
Represents the query input. It can contain either:

GoogleCloudDialogflowV2VoiceSelectionParams
Description of which voice to use for speech synthesis.

GoogleCloudDialogflowV2IntentMessageBasicCardButton
The button object that appears at the bottom of a card.

GoogleCloudDialogflowV2IntentTrainingPhrasePart
Represents a part of a training phrase.

GoogleCloudDialogflowV2Intent
Represents an intent. Intents convert a number of user expressions or patterns into an action. An action is an extraction of a user command or sentence semantics.

GoogleCloudDialogflowV2BatchCreateEntitiesRequest
The request message for EntityTypes.BatchCreateEntities.

GoogleCloudDialogflowV2IntentBatch
This message is a wrapper around a collection of intents.

GoogleCloudDialogflowV2IntentMessageCarouselSelect
The card for presenting a carousel of options to select from.

GoogleCloudDialogflowV2EntityTypeEntity
An entity entry for an associated entity type.

GoogleCloudDialogflowV2BatchDeleteIntentsRequest
The request message for Intents.BatchDeleteIntents.

GoogleCloudDialogflowV2SearchAgentsResponse
The response message for Agents.SearchAgents.

GoogleCloudDialogflowV2IntentParameter
Represents intent parameters.

GoogleCloudDialogflowV2IntentMessageQuickReplies
The quick replies response message.

GoogleCloudDialogflowV2IntentMessageCarouselSelectItem
An item in the carousel.

GoogleCloudDialogflowV2DetectIntentResponse
The message returned from the DetectIntent method.

GoogleCloudDialogflowV2BatchDeleteEntitiesRequest
The request message for EntityTypes.BatchDeleteEntities.

GoogleCloudDialogflowV2IntentMessageCardButton
Optional. Contains information about a button.

GoogleCloudDialogflowV2InputAudioConfig
Instructs the speech recognizer how to process the audio content.

GoogleCloudDialogflowV2SynthesizeSpeechConfig
Configuration of how speech should be synthesized.

GoogleCloudDialogflowV2BatchUpdateEntityTypesRequest
The request message for EntityTypes.BatchUpdateEntityTypes.

GoogleCloudDialogflowV2ImportAgentRequest
The request message for Agents.ImportAgent.

GoogleCloudDialogflowV2Context
Represents a context.

GoogleCloudDialogflowV2BatchUpdateEntitiesRequest
The request message for EntityTypes.BatchUpdateEntities.

GoogleCloudDialogflowV2RestoreAgentRequest
The request message for Agents.RestoreAgent.

GoogleCloudDialogflowV2IntentMessageSuggestion
The suggestion chip message that the user can tap to quickly post a reply to the conversation.

GoogleCloudDialogflowV2SessionEntityType
Represents a session entity type.

GoogleCloudDialogflowV2BatchUpdateIntentsRequest
The request message for Intents.BatchUpdateIntents.

GoogleCloudDialogflowV2SentimentAnalysisRequestConfig
Configures the types of sentiment analysis to perform.

GoogleCloudDialogflowV2ListEntityTypesResponse
The response message for EntityTypes.ListEntityTypes.

GoogleCloudDialogflowV2ListSessionEntityTypesResponse
The response message for SessionEntityTypes.ListSessionEntityTypes.

GoogleCloudDialogflowV2EventInput
Events allow for matching intents by event name instead of the natural language input. For instance, input <event: { name: "welcome_event", parameters: { name: "Sam" } }> can trigger a personalized welcome response. The parameter name may be used by the agent in the response: "Hello #welcome_event.name! What can I do for you today?".

GoogleCloudDialogflowV2ListContextsResponse
The response message for Contexts.ListContexts.

GoogleCloudDialogflowV2TextInput
Represents the natural language text to be processed.

GoogleCloudDialogflowV2IntentMessageCard
The card response message.

GoogleCloudDialogflowV2ExportAgentRequest
The request message for Agents.ExportAgent.

GoogleCloudDialogflowV2TrainAgentRequest
The request message for Agents.TrainAgent.

GoogleCloudDialogflowV2OutputAudioConfig
Instructs the speech synthesizer on how to generate the output audio content.

GoogleCloudDialogflowV2IntentMessageLinkOutSuggestion
The suggestion chip message that allows the user to jump out to the app or website associated with this agent.

GoogleCloudDialogflowV2EntityType
Represents an entity type. Entity types serve as a tool for extracting parameter values from natural language queries.

GoogleCloudDialogflowV2ListIntentsResponse
The response message for Intents.ListIntents.

GoogleCloudDialogflowV2SentimentAnalysisResult
The result of sentiment analysis as configured by sentiment_analysis_request_config.

GoogleCloudDialogflowV2IntentMessageListSelect
The card for presenting a list of options to select from.

GoogleCloudDialogflowV2IntentMessageSelectItemInfo
Additional info about the select item for when it is triggered in a dialog.

GoogleCloudDialogflowV2IntentMessageSuggestions
The collection of suggestions.

GoogleCloudDialogflowV2Sentiment
The sentiment, such as positive/negative feeling or association, for a unit of analysis, such as the query text.

GoogleCloudDialogflowV2QueryResult
Represents the result of conversational query or event processing.

GoogleCloudDialogflowV2IntentMessageBasicCard
The basic card message. Useful for displaying information.

GoogleCloudDialogflowV2IntentTrainingPhrase
Represents an example that the agent is trained on.

GoogleCloudDialogflowV2IntentMessageImage
The image response message.

GoogleCloudDialogflowV2QueryParameters
Represents the parameters of the conversational query.

GoogleCloudDialogflowV2IntentMessageSimpleResponse
The simple response message containing speech or text.

GoogleCloudDialogflowV2IntentMessageBasicCardButtonOpenUriAction
Opens the given URI.

GoogleCloudDialogflowV2Agent
Represents a conversational agent.

GoogleCloudDialogflowV2EntityTypeBatch
This message is a wrapper around a collection of entity types.

GoogleCloudDialogflowV2IntentMessageListSelectItem
An item in the list.

GoogleCloudDialogflowV2IntentFollowupIntentInfo
Represents a single followup intent in the chain.

GoogleCloudDialogflowV2BatchDeleteEntityTypesRequest
The request message for EntityTypes.BatchDeleteEntityTypes.

GoogleCloudDialogflowV2IntentMessageText
The text response message.

GoogleLongrunningOperation
This resource represents a long-running operation that is the result of a network API call.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

GoogleRpcStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details.

GoogleTypeLatLng
An object representing a latitude/longitude pair. This is expressed as a pair of doubles representing degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges.

JsonServerError
A utility type which can decode a server response that indicates error

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

ProjectAgentCall
Creates/updates the specified agent.

ProjectAgentEntityTypeBatchDeleteCall
Deletes entity types in the specified agent.

ProjectAgentEntityTypeBatchUpdateCall
Updates/Creates multiple entity types in the specified agent.

ProjectAgentEntityTypeCreateCall
Creates an entity type in the specified agent.

ProjectAgentEntityTypeDeleteCall
Deletes the specified entity type.

ProjectAgentEntityTypeEntityBatchCreateCall
Creates multiple new entities in the specified entity type.

ProjectAgentEntityTypeEntityBatchDeleteCall
Deletes entities in the specified entity type.

ProjectAgentEntityTypeEntityBatchUpdateCall
Updates or creates multiple entities in the specified entity type. This method does not affect entities in the entity type that aren't explicitly specified in the request.

ProjectAgentEntityTypeGetCall
Retrieves the specified entity type.

ProjectAgentEntityTypeListCall
Returns the list of all entity types in the specified agent.

ProjectAgentEntityTypePatchCall
Updates the specified entity type.

ProjectAgentExportCall
Exports the specified agent to a ZIP file.

ProjectAgentImportCall
Imports the specified agent from a ZIP file.

ProjectAgentIntentBatchDeleteCall
Deletes intents in the specified agent.

ProjectAgentIntentBatchUpdateCall
Updates/Creates multiple intents in the specified agent.

ProjectAgentIntentCreateCall
Creates an intent in the specified agent.

ProjectAgentIntentDeleteCall
Deletes the specified intent and its direct or indirect followup intents.

ProjectAgentIntentGetCall
Retrieves the specified intent.

ProjectAgentIntentListCall
Returns the list of all intents in the specified agent.

ProjectAgentIntentPatchCall
Updates the specified intent.

ProjectAgentRestoreCall
Restores the specified agent from a ZIP file.

ProjectAgentSearchCall
Returns the list of agents.

ProjectAgentSessionContextCreateCall
Creates a context.

ProjectAgentSessionContextDeleteCall
Deletes the specified context.

ProjectAgentSessionContextGetCall
Retrieves the specified context.

ProjectAgentSessionContextListCall
Returns the list of all contexts in the specified session.

ProjectAgentSessionContextPatchCall
Updates the specified context.

ProjectAgentSessionDeleteContextCall
Deletes all active contexts in the specified session.

ProjectAgentSessionDetectIntentCall
Processes a natural language query and returns structured, actionable data as a result. This method is not idempotent, because it may cause contexts and session entity types to be updated, which in turn might affect results of future queries.

ProjectAgentSessionEntityTypeCreateCall
Creates a session entity type.

ProjectAgentSessionEntityTypeDeleteCall
Deletes the specified session entity type.

ProjectAgentSessionEntityTypeGetCall
Retrieves the specified session entity type.

ProjectAgentSessionEntityTypeListCall
Returns the list of all session entity types in the specified session.

ProjectAgentSessionEntityTypePatchCall
Updates the specified session entity type.

ProjectAgentTrainCall
Trains the specified agent.

ProjectDeleteAgentCall
Deletes the specified agent.

ProjectGetAgentCall
Retrieves the specified agent.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Dialogflow hub.

ProjectOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

RangeResponseHeader

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ServerError

ServerMessage

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.