Module google_dialogflow2::api

Structs

Dialogflow
Central instance to access all Dialogflow related resource activities

GoogleCloudDialogflowV2Agent
A Dialogflow agent is a virtual agent that handles conversations with your end-users. It is a natural language understanding module that understands the nuances of human language. Dialogflow translates end-user text or audio during a conversation to structured data that your apps and services can understand. You design and build a Dialogflow agent to handle the types of conversations required for your system. For more information about agents, see the Agent guide.

GoogleCloudDialogflowV2AgentAssistantFeedback
Detail feedback of Agent Assist result.

GoogleCloudDialogflowV2AgentAssistantRecord
Represents a record of a human agent assist answer.

GoogleCloudDialogflowV2AnalyzeContentRequest
The request message for Participants.AnalyzeContent.

GoogleCloudDialogflowV2AnalyzeContentResponse
The response message for Participants.AnalyzeContent.

GoogleCloudDialogflowV2AnnotatedMessagePart
Represents a part of a message possibly annotated with an entity. The part can be an entity or purely a part of the message between two entities or message start/end.

GoogleCloudDialogflowV2AnswerFeedback
Represents feedback the customer has about the quality & correctness of a certain answer in a conversation.

GoogleCloudDialogflowV2AnswerRecord
Answer records are records to manage answer history and feedbacks for Dialogflow. Currently, answer record includes: - human agent assistant article suggestion - human agent assistant faq article It doesn't include: - DetectIntent intent matching - DetectIntent knowledge Answer records are not related to the conversation history in the Dialogflow Console. A Record is generated even when the end-user disables conversation history in the console. Records are created when there's a human agent assistant suggestion generated. A typical workflow for customers provide feedback to an answer is: 1. For human agent assistant, customers get suggestion via ListSuggestions API. Together with the answers, AnswerRecord.name are returned to the customers. 2. The customer uses the AnswerRecord.name to call the UpdateAnswerRecord method to send feedback about a specific answer that they believe is wrong.

GoogleCloudDialogflowV2ArticleAnswer
Represents article answer.

GoogleCloudDialogflowV2AutomatedAgentConfig
Defines the Automated Agent to connect to a conversation.

GoogleCloudDialogflowV2AutomatedAgentReply
Represents a response from an automated agent.

GoogleCloudDialogflowV2BatchCreateEntitiesRequest
The request message for EntityTypes.BatchCreateEntities.

GoogleCloudDialogflowV2BatchDeleteEntitiesRequest
The request message for EntityTypes.BatchDeleteEntities.

GoogleCloudDialogflowV2BatchDeleteEntityTypesRequest
The request message for EntityTypes.BatchDeleteEntityTypes.

GoogleCloudDialogflowV2BatchDeleteIntentsRequest
The request message for Intents.BatchDeleteIntents.

GoogleCloudDialogflowV2BatchUpdateEntitiesRequest
The request message for EntityTypes.BatchUpdateEntities.

GoogleCloudDialogflowV2BatchUpdateEntityTypesRequest
The request message for EntityTypes.BatchUpdateEntityTypes.

GoogleCloudDialogflowV2BatchUpdateIntentsRequest
There is no detailed description.

GoogleCloudDialogflowV2CompleteConversationRequest
The request message for Conversations.CompleteConversation.

GoogleCloudDialogflowV2Context
Dialogflow contexts are similar to natural language context. If a person says to you "they are orange", you need context in order to understand what "they" is referring to. Similarly, for Dialogflow to handle an end-user expression like that, it needs to be provided with context in order to correctly match an intent. Using contexts, you can control the flow of a conversation. You can configure contexts for an intent by setting input and output contexts, which are identified by string names. When an intent is matched, any configured output contexts for that intent become active. While any contexts are active, Dialogflow is more likely to match intents that are configured with input contexts that correspond to the currently active contexts. For more information about context, see the Contexts guide.

GoogleCloudDialogflowV2Conversation
Represents a conversation. A conversation is an interaction between an agent, including live agents and Dialogflow agents, and a support customer. Conversations can include phone calls and text-based chat sessions.

GoogleCloudDialogflowV2ConversationPhoneNumber
Represents a phone number for telephony integration. It allows for connecting a particular conversation over telephony.

GoogleCloudDialogflowV2ConversationProfile
Defines the services to connect to incoming Dialogflow conversations.

GoogleCloudDialogflowV2DetectIntentRequest
The request to detect user's intent.

GoogleCloudDialogflowV2DetectIntentResponse
The message returned from the DetectIntent method.

GoogleCloudDialogflowV2Document
A knowledge document to be used by a KnowledgeBase. For more information, see the knowledge base guide. Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents.

GoogleCloudDialogflowV2DocumentReloadStatus
The status of a reload attempt.

GoogleCloudDialogflowV2DtmfParameters
The message in the response that indicates the parameters of DTMF.

GoogleCloudDialogflowV2EntityType
Each intent parameter has a type, called the entity type, which dictates exactly how data from an end-user expression is extracted. Dialogflow provides predefined system entities that can match many common types of data. For example, there are system entities for matching dates, times, colors, email addresses, and so on. You can also create your own custom entities for matching custom data. For example, you could define a vegetable entity that can match the types of vegetables available for purchase with a grocery store agent. For more information, see the Entity guide.

GoogleCloudDialogflowV2EntityTypeBatch
This message is a wrapper around a collection of entity types.

GoogleCloudDialogflowV2EntityTypeEntity
An entity entry for an associated entity type.

GoogleCloudDialogflowV2Environment
You can create multiple versions of your agent and publish them to separate environments. When you edit an agent, you are editing the draft agent. At any point, you can save the draft agent as an agent version, which is an immutable snapshot of your agent. When you save the draft agent, it is published to the default environment. When you create agent versions, you can publish them to custom environments. You can create a variety of custom environments for: - testing - development - production - etc. For more information, see the versions and environments guide.

GoogleCloudDialogflowV2EventInput
Events allow for matching intents by event name instead of the natural language input. For instance, input `` can trigger a personalized welcome response. The parameter name may be used by the agent in the response: "Hello #welcome_event.name! What can I do for you today?".

GoogleCloudDialogflowV2ExportAgentRequest
The request message for Agents.ExportAgent.

GoogleCloudDialogflowV2FaqAnswer
Represents answer from "frequently asked questions".

GoogleCloudDialogflowV2Fulfillment
By default, your agent responds to a matched intent with a static response. As an alternative, you can provide a more dynamic response by using fulfillment. When you enable fulfillment for an intent, Dialogflow responds to that intent by calling a service that you define. For example, if an end-user wants to schedule a haircut on Friday, your service can check your database and respond to the end-user with availability information for Friday. For more information, see the fulfillment guide.

GoogleCloudDialogflowV2FulfillmentFeature
Whether fulfillment is enabled for the specific feature.

GoogleCloudDialogflowV2FulfillmentGenericWebService
Represents configuration for a generic web service. Dialogflow supports two mechanisms for authentications: - Basic authentication with username and password. - Authentication with additional authentication headers. More information could be found at: https://cloud.google.com/dialogflow/docs/fulfillment-configure.

GoogleCloudDialogflowV2HumanAgentAssistantConfig
Defines the Human Agent Assist to connect to a conversation.

GoogleCloudDialogflowV2HumanAgentAssistantConfigConversationModelConfig
Custom conversation models used in agent assist feature. Supported feature: ARTICLE_SUGGESTION, SMART_COMPOSE, SMART_REPLY.

GoogleCloudDialogflowV2HumanAgentAssistantConfigMessageAnalysisConfig
Configuration for analyses to run on each conversation message.

GoogleCloudDialogflowV2HumanAgentAssistantConfigSuggestionConfig
Detail human agent assistant config.

GoogleCloudDialogflowV2HumanAgentAssistantConfigSuggestionFeatureConfig
Config for suggestion features.

GoogleCloudDialogflowV2HumanAgentAssistantConfigSuggestionQueryConfig
Config for suggestion query.

GoogleCloudDialogflowV2HumanAgentAssistantConfigSuggestionQueryConfigContextFilterSettings
Settings that determine how to filter recent conversation context when generating suggestions.

GoogleCloudDialogflowV2HumanAgentAssistantConfigSuggestionQueryConfigDialogflowQuerySource
Dialogflow source setting. Supported feature: DIALOGFLOW_ASSIST.

GoogleCloudDialogflowV2HumanAgentAssistantConfigSuggestionQueryConfigDocumentQuerySource
Document source settings. Supported features: SMART_REPLY, SMART_COMPOSE.

GoogleCloudDialogflowV2HumanAgentAssistantConfigSuggestionQueryConfigKnowledgeBaseQuerySource
Knowledge base source settings. Supported features: ARTICLE_SUGGESTION, FAQ.

GoogleCloudDialogflowV2HumanAgentAssistantConfigSuggestionTriggerSettings
Settings of suggestion trigger.

GoogleCloudDialogflowV2HumanAgentHandoffConfig
Defines the hand off to a live agent, typically on which external agent service provider to connect to a conversation. Currently, this feature is not general available, please contact Google to get access.

GoogleCloudDialogflowV2HumanAgentHandoffConfigLivePersonConfig
Configuration specific to LivePerson (https://www.liveperson.com).

GoogleCloudDialogflowV2HumanAgentHandoffConfigSalesforceLiveAgentConfig
Configuration specific to Salesforce Live Agent.

GoogleCloudDialogflowV2ImportAgentRequest
The request message for Agents.ImportAgent.

GoogleCloudDialogflowV2InputAudioConfig
Instructs the speech recognizer how to process the audio content.

GoogleCloudDialogflowV2Intent
An intent categorizes an end-user's intention for one conversation turn. For each agent, you define many intents, where your combined intents can handle a complete conversation. When an end-user writes or says something, referred to as an end-user expression or end-user input, Dialogflow matches the end-user input to the best intent in your agent. Matching an intent is also known as intent classification. For more information, see the intent guide.

GoogleCloudDialogflowV2IntentBatch
This message is a wrapper around a collection of intents.

GoogleCloudDialogflowV2IntentFollowupIntentInfo
Represents a single followup intent in the chain.

GoogleCloudDialogflowV2IntentMessage
A rich response message. Corresponds to the intent Response field in the Dialogflow console. For more information, see Rich response messages.

GoogleCloudDialogflowV2IntentMessageBasicCard
The basic card message. Useful for displaying information.

GoogleCloudDialogflowV2IntentMessageBasicCardButton
The button object that appears at the bottom of a card.

GoogleCloudDialogflowV2IntentMessageBasicCardButtonOpenUriAction
Opens the given URI.

GoogleCloudDialogflowV2IntentMessageBrowseCarouselCard
Browse Carousel Card for Actions on Google. https://developers.google.com/actions/assistant/responses#browsing_carousel

GoogleCloudDialogflowV2IntentMessageBrowseCarouselCardBrowseCarouselCardItem
Browsing carousel tile

GoogleCloudDialogflowV2IntentMessageBrowseCarouselCardBrowseCarouselCardItemOpenUrlAction
Actions on Google action to open a given url.

GoogleCloudDialogflowV2IntentMessageCard
The card response message.

GoogleCloudDialogflowV2IntentMessageCardButton
Contains information about a button.

GoogleCloudDialogflowV2IntentMessageCarouselSelect
The card for presenting a carousel of options to select from.

GoogleCloudDialogflowV2IntentMessageCarouselSelectItem
An item in the carousel.

GoogleCloudDialogflowV2IntentMessageColumnProperties
Column properties for TableCard.

GoogleCloudDialogflowV2IntentMessageImage
The image response message.

GoogleCloudDialogflowV2IntentMessageLinkOutSuggestion
The suggestion chip message that allows the user to jump out to the app or website associated with this agent.

GoogleCloudDialogflowV2IntentMessageListSelect
The card for presenting a list of options to select from.

GoogleCloudDialogflowV2IntentMessageListSelectItem
An item in the list.

GoogleCloudDialogflowV2IntentMessageMediaContent
The media content card for Actions on Google.

GoogleCloudDialogflowV2IntentMessageMediaContentResponseMediaObject
Response media object for media content card.

GoogleCloudDialogflowV2IntentMessageQuickReplies
The quick replies response message.

GoogleCloudDialogflowV2IntentMessageSelectItemInfo
Additional info about the select item for when it is triggered in a dialog.

GoogleCloudDialogflowV2IntentMessageSimpleResponse
The simple response message containing speech or text.

GoogleCloudDialogflowV2IntentMessageSimpleResponses
The collection of simple response candidates. This message in QueryResult.fulfillment_messages and WebhookResponse.fulfillment_messages should contain only one SimpleResponse.

GoogleCloudDialogflowV2IntentMessageSuggestion
The suggestion chip message that the user can tap to quickly post a reply to the conversation.

GoogleCloudDialogflowV2IntentMessageSuggestions
The collection of suggestions.

GoogleCloudDialogflowV2IntentMessageTableCard
Table card for Actions on Google.

GoogleCloudDialogflowV2IntentMessageTableCardCell
Cell of TableCardRow.

GoogleCloudDialogflowV2IntentMessageTableCardRow
Row of TableCard.

GoogleCloudDialogflowV2IntentMessageText
The text response message.

GoogleCloudDialogflowV2IntentParameter
Represents intent parameters.

GoogleCloudDialogflowV2IntentTrainingPhrase
Represents an example that the agent is trained on.

GoogleCloudDialogflowV2IntentTrainingPhrasePart
Represents a part of a training phrase.

GoogleCloudDialogflowV2KnowledgeBase
A knowledge base represents a collection of knowledge documents that you provide to Dialogflow. Your knowledge documents contain information that may be useful during conversations with end-users. Some Dialogflow features use knowledge bases when looking for a response to an end-user input. For more information, see the knowledge base guide. Note: The projects.agent.knowledgeBases resource is deprecated; only use projects.knowledgeBases.

GoogleCloudDialogflowV2ListAnswerRecordsResponse
Response message for AnswerRecords.ListAnswerRecords.

GoogleCloudDialogflowV2ListContextsResponse
The response message for Contexts.ListContexts.

GoogleCloudDialogflowV2ListConversationProfilesResponse
The response message for ConversationProfiles.ListConversationProfiles.

GoogleCloudDialogflowV2ListConversationsResponse
The response message for Conversations.ListConversations.

GoogleCloudDialogflowV2ListDocumentsResponse
Response message for Documents.ListDocuments.

GoogleCloudDialogflowV2ListEntityTypesResponse
The response message for EntityTypes.ListEntityTypes.

GoogleCloudDialogflowV2ListEnvironmentsResponse
The response message for Environments.ListEnvironments.

GoogleCloudDialogflowV2ListIntentsResponse
The response message for Intents.ListIntents.

GoogleCloudDialogflowV2ListKnowledgeBasesResponse
Response message for KnowledgeBases.ListKnowledgeBases.

GoogleCloudDialogflowV2ListMessagesResponse
The response message for Conversations.ListMessages.

GoogleCloudDialogflowV2ListParticipantsResponse
The response message for Participants.ListParticipants.

GoogleCloudDialogflowV2ListSessionEntityTypesResponse
The response message for SessionEntityTypes.ListSessionEntityTypes.

GoogleCloudDialogflowV2LoggingConfig
Defines logging behavior for conversation lifecycle events.

GoogleCloudDialogflowV2Message
Represents a message posted into a conversation.

GoogleCloudDialogflowV2MessageAnnotation
Represents the result of annotation for the message.

GoogleCloudDialogflowV2NotificationConfig
Defines notification behavior.

GoogleCloudDialogflowV2OutputAudio
Represents the natural language speech audio to be played to the end user.

GoogleCloudDialogflowV2OutputAudioConfig
Instructs the speech synthesizer on how to generate the output audio content. If this audio config is supplied in a request, it overrides all existing text-to-speech settings applied to the agent.

GoogleCloudDialogflowV2Participant
Represents a conversation participant (human agent, virtual agent, end-user).

GoogleCloudDialogflowV2QueryInput
Represents the query input. It can contain either: 1. An audio config which instructs the speech recognizer how to process the speech audio. 2. A conversational query in the form of text,. 3. An event that specifies which intent to trigger.

GoogleCloudDialogflowV2QueryParameters
Represents the parameters of the conversational query.

GoogleCloudDialogflowV2QueryResult
Represents the result of conversational query or event processing.

GoogleCloudDialogflowV2ReloadDocumentRequest
Request message for Documents.ReloadDocument.

GoogleCloudDialogflowV2RestoreAgentRequest
The request message for Agents.RestoreAgent.

GoogleCloudDialogflowV2SearchAgentsResponse
The response message for Agents.SearchAgents.

GoogleCloudDialogflowV2Sentiment
The sentiment, such as positive/negative feeling or association, for a unit of analysis, such as the query text.

GoogleCloudDialogflowV2SentimentAnalysisRequestConfig
Configures the types of sentiment analysis to perform.

GoogleCloudDialogflowV2SentimentAnalysisResult
The result of sentiment analysis. Sentiment analysis inspects user input and identifies the prevailing subjective opinion, especially to determine a user's attitude as positive, negative, or neutral. For Participants.DetectIntent, it needs to be configured in DetectIntentRequest.query_params. For Participants.StreamingDetectIntent, it needs to be configured in StreamingDetectIntentRequest.query_params. And for Participants.AnalyzeContent and Participants.StreamingAnalyzeContent, it needs to be configured in ConversationProfile.human_agent_assistant_config

GoogleCloudDialogflowV2SessionEntityType
A session represents a conversation between a Dialogflow agent and an end-user. You can create special entities, called session entities, during a session. Session entities can extend or replace custom entity types and only exist during the session that they were created for. All session data, including session entities, is stored by Dialogflow for 20 minutes. For more information, see the session entity guide.

GoogleCloudDialogflowV2SpeechContext
Hints for the speech recognizer to help with recognition in a specific conversation state.

GoogleCloudDialogflowV2SpeechToTextConfig
Configures speech transcription for ConversationProfile.

GoogleCloudDialogflowV2SuggestArticlesRequest
The request message for Participants.SuggestArticles.

GoogleCloudDialogflowV2SuggestArticlesResponse
The response message for Participants.SuggestArticles.

GoogleCloudDialogflowV2SuggestFaqAnswersRequest
The request message for Participants.SuggestFaqAnswers.

GoogleCloudDialogflowV2SuggestFaqAnswersResponse
The request message for Participants.SuggestFaqAnswers.

GoogleCloudDialogflowV2SuggestionFeature
The type of Human Agent Assistant API suggestion to perform, and the maximum number of results to return for that type. Multiple Feature objects can be specified in the features list.

GoogleCloudDialogflowV2SuggestionResult
One response of different type of suggestion response which is used in the response of Participants.AnalyzeContent and Participants.AnalyzeContent, as well as HumanAgentAssistantEvent.

GoogleCloudDialogflowV2SynthesizeSpeechConfig
Configuration of how speech should be synthesized.

GoogleCloudDialogflowV2TextInput
Represents the natural language text to be processed.

GoogleCloudDialogflowV2TrainAgentRequest
The request message for Agents.TrainAgent.

GoogleCloudDialogflowV2ValidationError
Represents a single validation error.

GoogleCloudDialogflowV2ValidationResult
Represents the output of agent validation.

GoogleCloudDialogflowV2VoiceSelectionParams
Description of which voice to use for speech synthesis.

GoogleLongrunningListOperationsResponse
The response message for Operations.ListOperations.

GoogleLongrunningOperation
This resource represents a long-running operation that is the result of a network API call.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GoogleRpcStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

GoogleTypeLatLng
An object that represents a latitude/longitude pair. This is expressed as a pair of doubles to represent degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges.

ProjectAgentEntityTypeBatchDeleteCall
Deletes entity types in the specified agent. Operation

ProjectAgentEntityTypeBatchUpdateCall
Updates/Creates multiple entity types in the specified agent. Operation

ProjectAgentEntityTypeCreateCall
Creates an entity type in the specified agent.

ProjectAgentEntityTypeDeleteCall
Deletes the specified entity type.

ProjectAgentEntityTypeEntityBatchCreateCall
Creates multiple new entities in the specified entity type. Operation

ProjectAgentEntityTypeEntityBatchDeleteCall
Deletes entities in the specified entity type. Operation

ProjectAgentEntityTypeEntityBatchUpdateCall
Updates or creates multiple entities in the specified entity type. This method does not affect entities in the entity type that aren't explicitly specified in the request. Operation

ProjectAgentEntityTypeGetCall
Retrieves the specified entity type.

ProjectAgentEntityTypeListCall
Returns the list of all entity types in the specified agent.

ProjectAgentEntityTypePatchCall
Updates the specified entity type.

ProjectAgentEnvironmentIntentListCall
Returns the list of all intents in the specified agent.

ProjectAgentEnvironmentListCall
Returns the list of all non-draft environments of the specified agent.

ProjectAgentEnvironmentUserSessionContextCreateCall
Creates a context. If the specified context already exists, overrides the context.

ProjectAgentEnvironmentUserSessionContextDeleteCall
Deletes the specified context.

ProjectAgentEnvironmentUserSessionContextGetCall
Retrieves the specified context.

ProjectAgentEnvironmentUserSessionContextListCall
Returns the list of all contexts in the specified session.

ProjectAgentEnvironmentUserSessionContextPatchCall
Updates the specified context.

ProjectAgentEnvironmentUserSessionDeleteContextCall
Deletes all active contexts in the specified session.

ProjectAgentEnvironmentUserSessionDetectIntentCall
Processes a natural language query and returns structured, actionable data as a result. This method is not idempotent, because it may cause contexts and session entity types to be updated, which in turn might affect results of future queries. Note: Always use agent versions for production traffic. See Versions and environments.

ProjectAgentEnvironmentUserSessionEntityTypeCreateCall
Creates a session entity type. If the specified session entity type already exists, overrides the session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectAgentEnvironmentUserSessionEntityTypeDeleteCall
Deletes the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectAgentEnvironmentUserSessionEntityTypeGetCall
Retrieves the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectAgentEnvironmentUserSessionEntityTypeListCall
Returns the list of all session entity types in the specified session. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectAgentEnvironmentUserSessionEntityTypePatchCall
Updates the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectAgentExportCall
Exports the specified agent to a ZIP file. Operation

ProjectAgentGetFulfillmentCall
Retrieves the fulfillment.

ProjectAgentGetValidationResultCall
Gets agent validation result. Agent validation is performed during training time and is updated automatically when training is completed.

ProjectAgentImportCall
Imports the specified agent from a ZIP file. Uploads new intents and entity types without deleting the existing ones. Intents and entity types with the same name are replaced with the new versions from ImportAgentRequest. After the import, the imported draft agent will be trained automatically (unless disabled in agent settings). However, once the import is done, training may not be completed yet. Please call TrainAgent and wait for the operation it returns in order to train explicitly. Operation An operation which tracks when importing is complete. It only tracks when the draft agent is updated not when it is done training.

ProjectAgentIntentBatchDeleteCall
Deletes intents in the specified agent. Operation

ProjectAgentIntentBatchUpdateCall
Updates/Creates multiple intents in the specified agent. Operation

ProjectAgentIntentCreateCall
Creates an intent in the specified agent.

ProjectAgentIntentDeleteCall
Deletes the specified intent and its direct or indirect followup intents.

ProjectAgentIntentGetCall
Retrieves the specified intent.

ProjectAgentIntentListCall
Returns the list of all intents in the specified agent.

ProjectAgentIntentPatchCall
Updates the specified intent.

ProjectAgentKnowledgeBaseCreateCall
Creates a knowledge base.

ProjectAgentKnowledgeBaseDeleteCall
Deletes the specified knowledge base.

ProjectAgentKnowledgeBaseDocumentCreateCall
Creates a new document. Operation

ProjectAgentKnowledgeBaseDocumentDeleteCall
Deletes the specified document. Operation

ProjectAgentKnowledgeBaseDocumentGetCall
Retrieves the specified document.

ProjectAgentKnowledgeBaseDocumentListCall
Returns the list of all documents of the knowledge base.

ProjectAgentKnowledgeBaseDocumentPatchCall
Updates the specified document. Operation

ProjectAgentKnowledgeBaseDocumentReloadCall
Reloads the specified document from its specified source, content_uri or content. The previously loaded content of the document will be deleted. Note: Even when the content of the document has not changed, there still may be side effects because of internal implementation changes. Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents. Operation

ProjectAgentKnowledgeBaseGetCall
Retrieves the specified knowledge base.

ProjectAgentKnowledgeBaseListCall
Returns the list of all knowledge bases of the specified agent.

ProjectAgentKnowledgeBasePatchCall
Updates the specified knowledge base.

ProjectAgentRestoreCall
Restores the specified agent from a ZIP file. Replaces the current agent version with a new one. All the intents and entity types in the older version are deleted. After the restore, the restored draft agent will be trained automatically (unless disabled in agent settings). However, once the restore is done, training may not be completed yet. Please call TrainAgent and wait for the operation it returns in order to train explicitly. Operation An operation which tracks when restoring is complete. It only tracks when the draft agent is updated not when it is done training.

ProjectAgentSearchCall
Returns the list of agents. Since there is at most one conversational agent per project, this method is useful primarily for listing all agents across projects the caller has access to. One can achieve that with a wildcard project collection id "-". Refer to List Sub-Collections.

ProjectAgentSessionContextCreateCall
Creates a context. If the specified context already exists, overrides the context.

ProjectAgentSessionContextDeleteCall
Deletes the specified context.

ProjectAgentSessionContextGetCall
Retrieves the specified context.

ProjectAgentSessionContextListCall
Returns the list of all contexts in the specified session.

ProjectAgentSessionContextPatchCall
Updates the specified context.

ProjectAgentSessionDeleteContextCall
Deletes all active contexts in the specified session.

ProjectAgentSessionDetectIntentCall
Processes a natural language query and returns structured, actionable data as a result. This method is not idempotent, because it may cause contexts and session entity types to be updated, which in turn might affect results of future queries. Note: Always use agent versions for production traffic. See Versions and environments.

ProjectAgentSessionEntityTypeCreateCall
Creates a session entity type. If the specified session entity type already exists, overrides the session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectAgentSessionEntityTypeDeleteCall
Deletes the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectAgentSessionEntityTypeGetCall
Retrieves the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectAgentSessionEntityTypeListCall
Returns the list of all session entity types in the specified session. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectAgentSessionEntityTypePatchCall
Updates the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectAgentTrainCall
Trains the specified agent. Operation

ProjectAgentUpdateFulfillmentCall
Updates the fulfillment.

ProjectAnswerRecordListCall
Returns the list of all answer records in the specified project in reverse chronological order.

ProjectAnswerRecordPatchCall
Updates the specified answer record.

ProjectConversationCompleteCall
Completes the specified conversation. Finished conversations are purged from the database after 30 days.

ProjectConversationCreateCall
Creates a new conversation. Conversations are auto-completed after 24 hours. Conversation Lifecycle: There are two stages during a conversation: Automated Agent Stage and Assist Stage. For Automated Agent Stage, there will be a dialogflow agent responding to user queries. For Assist Stage, there's no dialogflow agent responding to user queries. But we will provide suggestions which are generated from conversation. If Conversation.conversation_profile is configured for a dialogflow agent, conversation will start from Automated Agent Stage, otherwise, it will start from Assist Stage. And during Automated Agent Stage, once an Intent with Intent.live_agent_handoff is triggered, conversation will transfer to Assist Stage.

ProjectConversationGetCall
Retrieves the specific conversation.

ProjectConversationListCall
Returns the list of all conversations in the specified project.

ProjectConversationMessageListCall
Lists messages that belong to a given conversation. messages are ordered by create_time in descending order. To fetch updates without duplication, send request with filter create_time_epoch_microseconds > [first item's create_time of previous request] and empty page_token.

ProjectConversationParticipantAnalyzeContentCall
Adds a text (chat, for example), or audio (phone recording, for example) message from a participant into the conversation. Note: Always use agent versions for production traffic sent to virtual agents. See Versions and environments.

ProjectConversationParticipantCreateCall
Creates a new participant in a conversation.

ProjectConversationParticipantGetCall
Retrieves a conversation participant.

ProjectConversationParticipantListCall
Returns the list of all participants in the specified conversation.

ProjectConversationParticipantPatchCall
Updates the specified participant.

ProjectConversationParticipantSuggestionSuggestArticleCall
Gets suggested articles for a participant based on specific historical messages.

ProjectConversationParticipantSuggestionSuggestFaqAnswerCall
Gets suggested faq answers for a participant based on specific historical messages.

ProjectConversationProfileCreateCall
Creates a conversation profile in the specified project. ConversationProfile.CreateTime and ConversationProfile.UpdateTime aren't populated in the response. You can retrieve them via GetConversationProfile API.

ProjectConversationProfileDeleteCall
Deletes the specified conversation profile.

ProjectConversationProfileGetCall
Retrieves the specified conversation profile.

ProjectConversationProfileListCall
Returns the list of all conversation profiles in the specified project.

ProjectConversationProfilePatchCall
Updates the specified conversation profile. ConversationProfile.CreateTime and ConversationProfile.UpdateTime aren't populated in the response. You can retrieve them via GetConversationProfile API.

ProjectDeleteAgentCall
Deletes the specified agent.

ProjectGetAgentCall
Retrieves the specified agent.

ProjectKnowledgeBaseCreateCall
Creates a knowledge base.

ProjectKnowledgeBaseDeleteCall
Deletes the specified knowledge base.

ProjectKnowledgeBaseDocumentCreateCall
Creates a new document. Operation

ProjectKnowledgeBaseDocumentDeleteCall
Deletes the specified document. Operation

ProjectKnowledgeBaseDocumentGetCall
Retrieves the specified document.

ProjectKnowledgeBaseDocumentListCall
Returns the list of all documents of the knowledge base.

ProjectKnowledgeBaseDocumentPatchCall
Updates the specified document. Operation

ProjectKnowledgeBaseDocumentReloadCall
Reloads the specified document from its specified source, content_uri or content. The previously loaded content of the document will be deleted. Note: Even when the content of the document has not changed, there still may be side effects because of internal implementation changes. Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents. Operation

ProjectKnowledgeBaseGetCall
Retrieves the specified knowledge base.

ProjectKnowledgeBaseListCall
Returns the list of all knowledge bases of the specified agent.

ProjectKnowledgeBasePatchCall
Updates the specified knowledge base.

ProjectLocationAgentEntityTypeBatchDeleteCall
Deletes entity types in the specified agent. Operation

ProjectLocationAgentEntityTypeBatchUpdateCall
Updates/Creates multiple entity types in the specified agent. Operation

ProjectLocationAgentEntityTypeCreateCall
Creates an entity type in the specified agent.

ProjectLocationAgentEntityTypeDeleteCall
Deletes the specified entity type.

ProjectLocationAgentEntityTypeEntityBatchCreateCall
Creates multiple new entities in the specified entity type. Operation

ProjectLocationAgentEntityTypeEntityBatchDeleteCall
Deletes entities in the specified entity type. Operation

ProjectLocationAgentEntityTypeEntityBatchUpdateCall
Updates or creates multiple entities in the specified entity type. This method does not affect entities in the entity type that aren't explicitly specified in the request. Operation

ProjectLocationAgentEntityTypeGetCall
Retrieves the specified entity type.

ProjectLocationAgentEntityTypeListCall
Returns the list of all entity types in the specified agent.

ProjectLocationAgentEntityTypePatchCall
Updates the specified entity type.

ProjectLocationAgentEnvironmentListCall
Returns the list of all non-draft environments of the specified agent.

ProjectLocationAgentEnvironmentUserSessionContextCreateCall
Creates a context. If the specified context already exists, overrides the context.

ProjectLocationAgentEnvironmentUserSessionContextDeleteCall
Deletes the specified context.

ProjectLocationAgentEnvironmentUserSessionContextGetCall
Retrieves the specified context.

ProjectLocationAgentEnvironmentUserSessionContextListCall
Returns the list of all contexts in the specified session.

ProjectLocationAgentEnvironmentUserSessionContextPatchCall
Updates the specified context.

ProjectLocationAgentEnvironmentUserSessionDeleteContextCall
Deletes all active contexts in the specified session.

ProjectLocationAgentEnvironmentUserSessionDetectIntentCall
Processes a natural language query and returns structured, actionable data as a result. This method is not idempotent, because it may cause contexts and session entity types to be updated, which in turn might affect results of future queries. Note: Always use agent versions for production traffic. See Versions and environments.

ProjectLocationAgentEnvironmentUserSessionEntityTypeCreateCall
Creates a session entity type. If the specified session entity type already exists, overrides the session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectLocationAgentEnvironmentUserSessionEntityTypeDeleteCall
Deletes the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectLocationAgentEnvironmentUserSessionEntityTypeGetCall
Retrieves the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectLocationAgentEnvironmentUserSessionEntityTypeListCall
Returns the list of all session entity types in the specified session. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectLocationAgentEnvironmentUserSessionEntityTypePatchCall
Updates the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectLocationAgentExportCall
Exports the specified agent to a ZIP file. Operation

ProjectLocationAgentGetFulfillmentCall
Retrieves the fulfillment.

ProjectLocationAgentGetValidationResultCall
Gets agent validation result. Agent validation is performed during training time and is updated automatically when training is completed.

ProjectLocationAgentImportCall
Imports the specified agent from a ZIP file. Uploads new intents and entity types without deleting the existing ones. Intents and entity types with the same name are replaced with the new versions from ImportAgentRequest. After the import, the imported draft agent will be trained automatically (unless disabled in agent settings). However, once the import is done, training may not be completed yet. Please call TrainAgent and wait for the operation it returns in order to train explicitly. Operation An operation which tracks when importing is complete. It only tracks when the draft agent is updated not when it is done training.

ProjectLocationAgentIntentBatchDeleteCall
Deletes intents in the specified agent. Operation

ProjectLocationAgentIntentBatchUpdateCall
Updates/Creates multiple intents in the specified agent. Operation

ProjectLocationAgentIntentCreateCall
Creates an intent in the specified agent.

ProjectLocationAgentIntentDeleteCall
Deletes the specified intent and its direct or indirect followup intents.

ProjectLocationAgentIntentGetCall
Retrieves the specified intent.

ProjectLocationAgentIntentListCall
Returns the list of all intents in the specified agent.

ProjectLocationAgentIntentPatchCall
Updates the specified intent.

ProjectLocationAgentRestoreCall
Restores the specified agent from a ZIP file. Replaces the current agent version with a new one. All the intents and entity types in the older version are deleted. After the restore, the restored draft agent will be trained automatically (unless disabled in agent settings). However, once the restore is done, training may not be completed yet. Please call TrainAgent and wait for the operation it returns in order to train explicitly. Operation An operation which tracks when restoring is complete. It only tracks when the draft agent is updated not when it is done training.

ProjectLocationAgentSearchCall
Returns the list of agents. Since there is at most one conversational agent per project, this method is useful primarily for listing all agents across projects the caller has access to. One can achieve that with a wildcard project collection id "-". Refer to List Sub-Collections.

ProjectLocationAgentSessionContextCreateCall
Creates a context. If the specified context already exists, overrides the context.

ProjectLocationAgentSessionContextDeleteCall
Deletes the specified context.

ProjectLocationAgentSessionContextGetCall
Retrieves the specified context.

ProjectLocationAgentSessionContextListCall
Returns the list of all contexts in the specified session.

ProjectLocationAgentSessionContextPatchCall
Updates the specified context.

ProjectLocationAgentSessionDeleteContextCall
Deletes all active contexts in the specified session.

ProjectLocationAgentSessionDetectIntentCall
Processes a natural language query and returns structured, actionable data as a result. This method is not idempotent, because it may cause contexts and session entity types to be updated, which in turn might affect results of future queries. Note: Always use agent versions for production traffic. See Versions and environments.

ProjectLocationAgentSessionEntityTypeCreateCall
Creates a session entity type. If the specified session entity type already exists, overrides the session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectLocationAgentSessionEntityTypeDeleteCall
Deletes the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectLocationAgentSessionEntityTypeGetCall
Retrieves the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectLocationAgentSessionEntityTypeListCall
Returns the list of all session entity types in the specified session. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectLocationAgentSessionEntityTypePatchCall
Updates the specified session entity type. This method doesn't work with Google Assistant integration. Contact Dialogflow support if you need to use session entities with Google Assistant integration.

ProjectLocationAgentTrainCall
Trains the specified agent. Operation

ProjectLocationAgentUpdateFulfillmentCall
Updates the fulfillment.

ProjectLocationAnswerRecordListCall
Returns the list of all answer records in the specified project in reverse chronological order.

ProjectLocationAnswerRecordPatchCall
Updates the specified answer record.

ProjectLocationConversationCompleteCall
Completes the specified conversation. Finished conversations are purged from the database after 30 days.

ProjectLocationConversationCreateCall
Creates a new conversation. Conversations are auto-completed after 24 hours. Conversation Lifecycle: There are two stages during a conversation: Automated Agent Stage and Assist Stage. For Automated Agent Stage, there will be a dialogflow agent responding to user queries. For Assist Stage, there's no dialogflow agent responding to user queries. But we will provide suggestions which are generated from conversation. If Conversation.conversation_profile is configured for a dialogflow agent, conversation will start from Automated Agent Stage, otherwise, it will start from Assist Stage. And during Automated Agent Stage, once an Intent with Intent.live_agent_handoff is triggered, conversation will transfer to Assist Stage.

ProjectLocationConversationGetCall
Retrieves the specific conversation.

ProjectLocationConversationListCall
Returns the list of all conversations in the specified project.

ProjectLocationConversationMessageListCall
Lists messages that belong to a given conversation. messages are ordered by create_time in descending order. To fetch updates without duplication, send request with filter create_time_epoch_microseconds > [first item's create_time of previous request] and empty page_token.

ProjectLocationConversationParticipantAnalyzeContentCall
Adds a text (chat, for example), or audio (phone recording, for example) message from a participant into the conversation. Note: Always use agent versions for production traffic sent to virtual agents. See Versions and environments.

ProjectLocationConversationParticipantCreateCall
Creates a new participant in a conversation.

ProjectLocationConversationParticipantGetCall
Retrieves a conversation participant.

ProjectLocationConversationParticipantListCall
Returns the list of all participants in the specified conversation.

ProjectLocationConversationParticipantPatchCall
Updates the specified participant.

ProjectLocationConversationParticipantSuggestionSuggestArticleCall
Gets suggested articles for a participant based on specific historical messages.

ProjectLocationConversationParticipantSuggestionSuggestFaqAnswerCall
Gets suggested faq answers for a participant based on specific historical messages.

ProjectLocationConversationProfileCreateCall
Creates a conversation profile in the specified project. ConversationProfile.CreateTime and ConversationProfile.UpdateTime aren't populated in the response. You can retrieve them via GetConversationProfile API.

ProjectLocationConversationProfileDeleteCall
Deletes the specified conversation profile.

ProjectLocationConversationProfileGetCall
Retrieves the specified conversation profile.

ProjectLocationConversationProfileListCall
Returns the list of all conversation profiles in the specified project.

ProjectLocationConversationProfilePatchCall
Updates the specified conversation profile. ConversationProfile.CreateTime and ConversationProfile.UpdateTime aren't populated in the response. You can retrieve them via GetConversationProfile API.

ProjectLocationDeleteAgentCall
Deletes the specified agent.

ProjectLocationGetAgentCall
Retrieves the specified agent.

ProjectLocationKnowledgeBaseCreateCall
Creates a knowledge base.

ProjectLocationKnowledgeBaseDeleteCall
Deletes the specified knowledge base.

ProjectLocationKnowledgeBaseDocumentCreateCall
Creates a new document. Operation

ProjectLocationKnowledgeBaseDocumentDeleteCall
Deletes the specified document. Operation

ProjectLocationKnowledgeBaseDocumentGetCall
Retrieves the specified document.

ProjectLocationKnowledgeBaseDocumentListCall
Returns the list of all documents of the knowledge base.

ProjectLocationKnowledgeBaseDocumentPatchCall
Updates the specified document. Operation

ProjectLocationKnowledgeBaseDocumentReloadCall
Reloads the specified document from its specified source, content_uri or content. The previously loaded content of the document will be deleted. Note: Even when the content of the document has not changed, there still may be side effects because of internal implementation changes. Note: The projects.agent.knowledgeBases.documents resource is deprecated; only use projects.knowledgeBases.documents. Operation

ProjectLocationKnowledgeBaseGetCall
Retrieves the specified knowledge base.

ProjectLocationKnowledgeBaseListCall
Returns the list of all knowledge bases of the specified agent.

ProjectLocationKnowledgeBasePatchCall
Updates the specified knowledge base.

ProjectLocationOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectLocationSetAgentCall
Creates/updates the specified agent.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Dialogflow hub.

ProjectOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectSetAgentCall
Creates/updates the specified agent.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.