Struct google_dialogflow2::api::GoogleCloudDialogflowV2ConversationProfile

pub struct GoogleCloudDialogflowV2ConversationProfile {
    pub automated_agent_config: Option<GoogleCloudDialogflowV2AutomatedAgentConfig>,
    pub create_time: Option<String>,
    pub display_name: Option<String>,
    pub human_agent_assistant_config: Option<GoogleCloudDialogflowV2HumanAgentAssistantConfig>,
    pub human_agent_handoff_config: Option<GoogleCloudDialogflowV2HumanAgentHandoffConfig>,
    pub language_code: Option<String>,
    pub logging_config: Option<GoogleCloudDialogflowV2LoggingConfig>,
    pub name: Option<String>,
    pub new_message_event_notification_config: Option<GoogleCloudDialogflowV2NotificationConfig>,
    pub notification_config: Option<GoogleCloudDialogflowV2NotificationConfig>,
    pub stt_config: Option<GoogleCloudDialogflowV2SpeechToTextConfig>,
    pub update_time: Option<String>,
}

Defines the services to connect to incoming Dialogflow conversations.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

conversation profiles create projects (request|response)

conversation profiles get projects (response)

conversation profiles patch projects (request|response)

locations conversation profiles create projects (request|response)

locations conversation profiles get projects (response)

locations conversation profiles patch projects (request|response)

Fields
automated_agent_config: Option<GoogleCloudDialogflowV2AutomatedAgentConfig>

Configuration for an automated agent to use with this profile.

create_time: Option<String>

Output only. Create time of the conversation profile.

display_name: Option<String>

Required. Human readable name for this profile. Max length 1024 bytes.

human_agent_assistant_config: Option<GoogleCloudDialogflowV2HumanAgentAssistantConfig>

Configuration for agent assistance to use with this profile.

human_agent_handoff_config: Option<GoogleCloudDialogflowV2HumanAgentHandoffConfig>

Configuration for connecting to a live agent. Currently, this feature is not general available, please contact Google to get access.

language_code: Option<String>

Language which represents the conversationProfile. If unspecified, the default language code en-us applies. Users need to create a ConversationProfile for each language they want to support.

logging_config: Option<GoogleCloudDialogflowV2LoggingConfig>

Configuration for logging conversation lifecycle events.

name: Option<String>

The unique identifier of this conversation profile. Format: projects//locations//conversationProfiles/.

new_message_event_notification_config: Option<GoogleCloudDialogflowV2NotificationConfig>

Configuration for publishing new message events. Event will be sent in format of ConversationEvent

notification_config: Option<GoogleCloudDialogflowV2NotificationConfig>

Configuration for publishing conversation lifecycle events.

stt_config: Option<GoogleCloudDialogflowV2SpeechToTextConfig>

Settings for speech transcription.

update_time: Option<String>

Output only. Update time of the conversation profile.