Struct google_dialogflow2::GoogleCloudDialogflowV2Intent

pub struct GoogleCloudDialogflowV2Intent {
    pub webhook_state: Option<String>,
    pub default_response_platforms: Option<Vec<String>>,
    pub action: Option<String>,
    pub output_contexts: Option<Vec<GoogleCloudDialogflowV2Context>>,
    pub is_fallback: Option<bool>,
    pub ml_disabled: Option<bool>,
    pub display_name: Option<String>,
    pub name: Option<String>,
    pub parameters: Option<Vec<GoogleCloudDialogflowV2IntentParameter>>,
    pub parent_followup_intent_name: Option<String>,
    pub followup_intent_info: Option<Vec<GoogleCloudDialogflowV2IntentFollowupIntentInfo>>,
    pub training_phrases: Option<Vec<GoogleCloudDialogflowV2IntentTrainingPhrase>>,
    pub messages: Option<Vec<GoogleCloudDialogflowV2IntentMessage>>,
    pub priority: Option<i32>,
    pub root_followup_intent_name: Option<String>,
    pub input_context_names: Option<Vec<String>>,
    pub reset_contexts: Option<bool>,
    pub events: Option<Vec<String>>,
}

Represents an intent. Intents convert a number of user expressions or patterns into an action. An action is an extraction of a user command or sentence semantics.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

agent intents get projects (response)

agent intents create projects (request|response)

agent intents patch projects (request|response)

Fields
webhook_state: Option<String>

Optional. Indicates whether webhooks are enabled for the intent.

default_response_platforms: Option<Vec<String>>

Optional. The list of platforms for which the first responses will be copied from the messages in PLATFORM_UNSPECIFIED (i.e. default platform).

action: Option<String>

Optional. The name of the action associated with the intent. Note: The action name must not contain whitespaces.

output_contexts: Option<Vec<GoogleCloudDialogflowV2Context>>

Optional. The collection of contexts that are activated when the intent is matched. Context messages in this collection should not set the parameters field. Setting the lifespan_count to 0 will reset the context when the intent is matched. Format: projects/<Project ID>/agent/sessions/-/contexts/<Context ID>.

is_fallback: Option<bool>

Optional. Indicates whether this is a fallback intent.

ml_disabled: Option<bool>

Optional. Indicates whether Machine Learning is disabled for the intent. Note: If ml_diabled setting is set to true, then this intent is not taken into account during inference in ML ONLY match mode. Also, auto-markup in the UI is turned off.

display_name: Option<String>

Required. The name of this intent.

name: Option<String>

The unique identifier of this intent. Required for Intents.UpdateIntent and Intents.BatchUpdateIntents methods. Format: projects/<Project ID>/agent/intents/<Intent ID>.

parameters: Option<Vec<GoogleCloudDialogflowV2IntentParameter>>

Optional. The collection of parameters associated with the intent.

parent_followup_intent_name: Option<String>

Read-only after creation. The unique identifier of the parent intent in the chain of followup intents. You can set this field when creating an intent, for example with CreateIntent or BatchUpdateIntents, in order to make this intent a followup intent.

It identifies the parent followup intent. Format: projects/<Project ID>/agent/intents/<Intent ID>.

followup_intent_info: Option<Vec<GoogleCloudDialogflowV2IntentFollowupIntentInfo>>

Read-only. Information about all followup intents that have this intent as a direct or indirect parent. We populate this field only in the output.

training_phrases: Option<Vec<GoogleCloudDialogflowV2IntentTrainingPhrase>>

Optional. The collection of examples that the agent is trained on.

messages: Option<Vec<GoogleCloudDialogflowV2IntentMessage>>

Optional. The collection of rich messages corresponding to the Response field in the Dialogflow console.

priority: Option<i32>

Optional. The priority of this intent. Higher numbers represent higher priorities. If this is zero or unspecified, we use the default priority 500000.

Negative numbers mean that the intent is disabled.

root_followup_intent_name: Option<String>

Read-only. The unique identifier of the root intent in the chain of followup intents. It identifies the correct followup intents chain for this intent. We populate this field only in the output.

Format: projects/<Project ID>/agent/intents/<Intent ID>.

input_context_names: Option<Vec<String>>

Optional. The list of context names required for this intent to be triggered. Format: projects/<Project ID>/agent/sessions/-/contexts/<Context ID>.

reset_contexts: Option<bool>

Optional. Indicates whether to delete all contexts in the current session when this intent is matched.

events: Option<Vec<String>>

Optional. The collection of event names that trigger the intent. If the collection of input contexts is not empty, all of the contexts must be present in the active user session for an event to trigger this intent.